backend/python/agents/data_analysis_agent/requirements.txt
grpcio
grpcio-tools
python3-etcd3
pandas
numpy
protobuf
matplotlib